requirements.txt
streamlit 
google-generativeai
python-dotenv
langchain-community
faiss-cpu
langchain-google-genai
sentence-transformers
python-dotenv